+++ react/components/SKUSelector/components/Variation.tsx
@@ -5,7 +5,7 @@ import { SliderLayout } from 'vtex.slider-layout'
 import { findIndex, propEq } from 'ramda'
 import classnames from 'classnames'
 import { useProduct } from 'vtex.product-context'
-import { ResponsiveInput } from 'vtex.responsive-values'
+import { ResponsiveValuesTypes } from 'vtex.responsive-values'
 
 import { stripUrl, isColor, slug } from '../utils'
 import styles from '../styles.css'
@@ -14,6 +14,7 @@ import { imageUrlForSize, VARIATION_IMG_SIZE } from '../../module/images'
 import ErrorMessage from './ErrorMessage'
 import SelectModeVariation from './SelectVariationMode'
 import SelectorItem from './SelectorItem'
+import { ShowVariationsLabels } from './SKUSelector'
 
 interface Props {
   variation: DisplayVariation
@@ -25,13 +26,13 @@ interface Props {
   imageHeight?: number
   imageWidth?: number
   showBorders?: boolean
-  showLabel: boolean
+  showLabel: ShowVariationsLabels
   containerClasses?: string
   showErrorMessage: boolean
   mode?: string
   sliderDisplayThreshold: number
   sliderArrowSize: number
-  sliderItemsPerPage: ResponsiveInput<number>
+  sliderItemsPerPage: ResponsiveValuesTypes.ResponsiveValue<number>
 }
 
 const ITEMS_VISIBLE_THRESHOLD = 2
@@ -106,6 +107,18 @@ const Variation: FC<Props> = ({
     fullWidth: false,
   }
 
+  // The following code is here to maintain backwards compatibility
+  let variationLabel = ''
+
+  if (typeof showLabel === 'boolean') {
+    variationLabel = showLabel ? 'variation' : 'none'
+  } else {
+    variationLabel = showLabel
+  }
+
+  const showVariationLabelName =
+    variationLabel === 'variation' || variationLabel === 'variationAndItemValue'
+
   const selectorItemsArray = displayOptions.map(option => {
     return (
       <SelectorItem
@@ -126,6 +139,8 @@ const Variation: FC<Props> = ({
         }
         imageLabel={option.image?.imageLabel}
         isImpossible={option.impossible}
+        variationLabel={variationLabel}
+        label={name}
       />
     )
   })
@@ -134,7 +149,7 @@ const Variation: FC<Props> = ({
     <div className={containerClasses}>
       <div className={`${styles.skuSelectorNameContainer} ma1`}>
         <div className={`${styles.skuSelectorTextContainer} db mb3`}>
-          {showLabel && (
+          {showVariationLabelName && (
             <span
               className={`${styles.skuSelectorName} c-muted-1 t-small overflow-hidden`}
             >