+++ react/components/SearchBar/SearchBar.tsx
@@ -111,6 +111,7 @@ function SearchBar({
     string | undefined
   >()
 
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   const debouncedSetSearchTerm = useCallback(
     debounce((newValue: string) => {
       setSearchTerm(newValue)
@@ -251,13 +252,10 @@ function SearchBar({
             openMenu,
           }) => (
             <div
-              className={classNames(
-                'relative-m w-100',
-                withModifiers('searchBarInnerContainer', [
-                  isOpen ? 'opened' : '',
-                  inputValue ? 'filled' : '',
-                ])
-              )}
+              className={`relative-m w-100 ${withModifiers(
+                'searchBarInnerContainer',
+                [isOpen ? 'opened' : '', inputValue ? 'filled' : '']
+              )}`}
             >
               <AutocompleteInput
                 // eslint-disable-next-line jsx-a11y/no-autofocus